fix test injection and accepting long

+++ src/main/java/com/example/demo/calc/entity/CalculationResult.java
@@ -3,5 +3,5 @@ package com.example.demo.calc.entity;
 import com.example.demo.PojaGenerated;
 
 @PojaGenerated
-public record CalculationResult(double result) {
+public record CalculationResult(long result) {
 }
