+++ src/main/java/com/example/demo/calc/service/CalculationService.java
@@ -8,25 +8,25 @@ import org.springframework.stereotype.Component;
 @Component
 public class CalculationService {
 
-    public CalculationResult add(double a, double b) {
+    public CalculationResult add(long a, long b) {
         if (a <= 0 || b <= 0) {
             throw new ArithmeticException("Both numbers must be positive");
         }
         return new CalculationResult(a + b);
     }
 
-    public CalculationResult subtract(double a, double b) {
+    public CalculationResult subtract(long a, long b) {
 
         return new CalculationResult(a - b);
     }
 
-    public CalculationResult multiply(double a, double b) {
+    public CalculationResult multiply(long a, long b) {
 
         return new CalculationResult(a * b);
     }
 
-    public CalculationResult divide(double a, double b) {
-        if (b == 0) {
+    public CalculationResult divide(long a, long b) {
+        if (b == 0 ) {
             throw new ArithmeticException("Division by zero is not allowed");
         }
         return new CalculationResult(a / b);
